requirements.txt
Flask==3.1.0
Pillow==11.1.0
torch==2.6.0+cu118
torchvision==0.21.0+cu118
matplotlib==3.10.0
scikit-learn==1.6.1